# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Listen Together API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnonymousLoginRequest {
  displayName?: string;
}

export interface UserSession {
  userId: string;
  displayName: string;
  avatarColor: string;
}

export interface CreateRoomRequest {
  name: string;
  hostUserId: string;
  hostDisplayName: string;
}

export interface Track {
  videoId: string;
  title: string;
  thumbnail: string;
  duration: number;
}

export interface Room {
  id: string;
  name: string;
  inviteCode: string;
  hostUserId: string;
  memberCount: number;
  currentTrack?: Track;
  isPlaying: boolean;
  currentTime: number;
}

export interface ErrorResponse {
  error: string;
}