# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Listen Together API
 * OpenAPI spec version: 0.1.0
 */
import type { Track } from "./track";

export interface Room {
  id: string;
  name: string;
  inviteCode: string;
  hostUserId: string;
  memberCount: number;
  currentTrack?: Track;
  isPlaying: boolean;
  currentTime: number;
}