# lib/api-zod/src/generated/types/userSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Listen Together API
 * OpenAPI spec version: 0.1.0
 */

export interface UserSession {
  userId: string;
  displayName: string;
  avatarColor: string;
}